Refactor code for service-based API
This API is modeled off the great work by the guys over at
thenewstack.io. Read the article that they have there to get a
good understanding of what is going on here.

https://thenewstack.io/make-a-restful-json-api-go/

But to summarize, I'm keeping the App struct as the main entry
point. I'm utilizing Gin as my HTTP router since it's more popular
and even faster than gorilla/mux.

The point where I differ however is that I'm utilizing interfaces
so I can have a set of default route handlers for each of my models.

The general flow goes like this:
    Client -> Server -> Router -> Service -> Model -> Resource

A Service represents all the endpoints exposed by a Model:
 - GET
 - FETCH
 - CREATE
 - UPDATE
 - DELETE

Then the Model will fetch the required data from the database and
return in the form of a Resource, which I plan to be a JSON
document that will returned to the Client.

Note: I have not finished the default model handlers yet, so this
code isn't production ready quite yet.